# src/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/years": {
      "post": {
        "operationId": "YearsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data to create new city",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateYearDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "year": 2020
                  }
                },
                "multiple": {
                  "value": [
                    {
                      "year": 2020
                    },
                    {
                      "year": 2021
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a year",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YearDto"
                },
                "examples": {
                  "single": {
                    "summary": "Year DTO",
                    "value": {
                      "id": 1,
                      "year": 2020
                    }
                  },
                  "multiple": {
                    "summary": "Count",
                    "value": {
                      "count": 10
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create new year.",
        "tags": [
          "years"
        ]
      },
      "get": {
        "operationId": "YearsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved years.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/YearDto"
                  }
                },
                "examples": {
                  "application/json": {
                    "summary": "Year DTO array",
                    "value": [
                      {
                        "id": 1,
                        "year": 2020
                      },
                      {
                        "id": 2,
                        "year": 2021
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Return all years.",
        "tags": [
          "years"
        ]
      }
    },
    "/years/{id}": {
      "get": {
        "operationId": "YearsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved year by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YearDto"
                },
                "examples": {
                  "application/json": {
                    "summary": "Year DTO",
                    "value": {
                      "id": 1,
                      "year": 2020
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Return year by id.",
        "tags": [
          "years"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateYearDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "description": "The year number"
          }
        },
        "required": [
          "year"
        ]
      },
      "YearDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the year"
          },
          "year": {
            "type": "number",
            "description": "The year number"
          }
        },
        "required": [
          "id",
          "year"
        ]
      },
...
    }
  }
}